feat: CommandPalette + SearchPanel 解耦与视觉精修 (#1193)
- CommandPalette.tsx: 793→297行，提取 commandPaletteTypes.ts / commandPaletteHelpers.tsx / commandPaletteCommands.tsx
- SearchPanel.tsx: 973→217行，提取 searchPanelTypes.ts / SearchResultItems.tsx / SearchResultsArea.tsx / SearchPanelParts.tsx
- AC-1: 分组 section header uppercase + 分隔线
- AC-4: 命令图标颜色编码（suggestions=蓝/layout=黄/document=绿/project=蓝）
- 所有提取文件 ≤300 行
- 2353 测试全绿，typecheck 通过，lint 0 errors
